aaaa

+++ backend/src/main/java/com/trtl88/backend/models/Publisher.java
@@ -1,11 +1,11 @@
 package com.trtl88.backend.models;
 public class Publisher {
-    private long publisherId;
+    private int publisherId;
     private String name;
     private String address;
     private String phoneNumber;
 
-    public Publisher(String name, String address, String phoneNumber, long publisherId) {
+    public Publisher(String name, String address, String phoneNumber, int publisherId) {
         this.name = name;
         this.address = address;
         this.phoneNumber = phoneNumber;
@@ -37,10 +37,10 @@ public class Publisher {
     public void setPhoneNumber(String phoneNumber) {
         this.phoneNumber = phoneNumber;
     }
-    public long getPublisherId() {
+    public int getPublisherId() {
         return publisherId;
     }
-    public void setPublisherId(long publisherId) {
+    public void setPublisherId(int publisherId) {
         this.publisherId = publisherId;
     }
 }
